+++ examples/initialize-flex.ts
@@ -4,24 +4,23 @@ import { Flex, FlexConfig } from "../flex";
 
 TonClient.useBinaryLibrary(libNode);
 
- (async () => {
-    try {
-        // Flex Initialization
-        const FLEX_CONFIG: Partial<FlexConfig> = {
-            evr: {
-                sdk: {
-                    network: {
-                        endpoints: ["https://test.flex.everos.dev/graphql"],
-                    },
+(async () => {
+    // Flex Initialization
+    const FLEX_CONFIG: Partial<FlexConfig> = {
+        evr: {
+            sdk: {
+                network: {
+                    endpoints: ["https://test.flex.everos.dev/graphql"],
                 },
             },
-            superRoot: "Super Root address",
-        };
-        const flex = new Flex(FLEX_CONFIG);
-
+        },
+        superRoot: "Super Root address",
+    };
+    const flex = new Flex(FLEX_CONFIG);
+    try {
         await flex.close();
     } catch (err) {
-        console.error(err);
+        flex.evr.log.error(err);
         process.exit(1);
     }
 })();