+++ examples/make-order.ts
@@ -1,32 +1,39 @@
-import { Flex, Trader } from "../flex";
+import { Flex, makeOrderFinalized, Trader } from "../flex";
 import { CONFIG, EXAMPLES_FLEX_CONFIG } from "./examples";
+import { waitForMakeOrder } from "../flex/trader/make-order";
 //import { LogLevel } from "../contracts/helpers";
 
 (async () => {
+    const flex = new Flex(EXAMPLES_FLEX_CONFIG);
     try {
-        const flex = new Flex(EXAMPLES_FLEX_CONFIG);
         // flex.evr.log.level = LogLevel.DEBUG;
         const clientAddress = CONFIG.trader.client;
         const traderId = CONFIG.trader.id;
         const marketAddress = CONFIG.market;
 
-        let orderInfo = await Trader.makeOrder(flex, {
+        let result = await Trader.makeOrder(flex, {
             clientAddress: clientAddress,
             trader: {
                 id: traderId,
-                signer: "trader_1",
+                signer: "traderSigner",
             },
-            sell: true,
+            sell: false,
             marketAddress: marketAddress,
-            price: { units: 250000 },
-            amount: { units: 5000 },
+            price: { tokens: 20 },
+            amount: { tokens: 10 },
             waitForOrderbookUpdate: true,
+            
         });
-        console.log(`Order info`, JSON.stringify(orderInfo, undefined, "   "));
+        flex.evr.log.info("MakeOrder Initialization result on wallet", result);
+
+        if (!makeOrderFinalized(result)) {
+            result = await waitForMakeOrder(flex, result);
+            flex.evr.log.info("Finalized Make order result in orderbook", result);
+        }
 
         await flex.close();
     } catch (err) {
-        console.error(err);
+        flex.evr.log.error(err);
         process.exit(1);
     }
 })();